refactor: drop the one-time legacy-upgrade import (cuts beta.69)
Pre-1.0 there are no production installs to migrate, so the code that
imported a pre-SQLite install's config.json / device-labels.json /
localStorage into the SQLite store on first upgrade is removed:

- server: importLegacy + importConfigJson + importDeviceLabels + the
  legacyDeviceLabels reader; Db.getInstance no longer runs an import.
- client: migrateLocalStorage; index.ts boot keeps theme + cache warming
  via applyStoredTheme (drops the migrate call).
- ThemeToggle seeds the OS theme on a fresh install (the migration-flag
  gate / shouldSeedTheme is gone).
- GLOBAL_KEYS (runtime config schema, not migration) relocated to
  db/constants.ts; the SQLite store + all per-user settings/labels/
  accounts are unchanged.
- removed 4 import tests + ThemeToggle.test (only tested the gate helper);
  updated db.test + Config.test (the in-memory port->webPort alias stays;
  the removed config-trim assertions dropped).
- docs: removed smoke Module 17 (the migration test) from all three smoke
  docs; bumped the smoke target to beta.69.

release:beta. Fresh installs unaffected; a pre-store dev build's settings
won't carry across the upgrade (acceptable - no production installs predate
the store). CHANGELOG note under [Unreleased].

Verified: tsc 0, biome 0, webpack build OK, full suite green (1477).
