+++ src/app/client/ThemeToggle.ts
@@ -1,56 +1,29 @@
 import { firstPaintTheme, getTheme, notifyThemeChanged, setTheme } from '../public/themeEmbed';
-import { LEGACY_KEYS } from './migrateLocalStorage';
 import { settingsService } from './SettingsService';
 
 const MOON_SVG = `<svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="currentColor"><path d="M21.752 15.002A9.718 9.718 0 0112.478 3.002a9.72 9.72 0 00-7.557 11.263A9.72 9.72 0 0016.49 21.78a9.718 9.718 0 005.262-6.778z"/></svg>`;
 const SUN_SVG = `<svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="currentColor"><path d="M12 2a1 1 0 011 1v1a1 1 0 11-2 0V3a1 1 0 011-1zm0 16a1 1 0 011 1v1a1 1 0 11-2 0v-1a1 1 0 011-1zm8.66-12.66a1 1 0 010 1.414l-.707.707a1 1 0 11-1.414-1.414l.707-.707a1 1 0 011.414 0zM6.046 17.246a1 1 0 010 1.414l-.707.707a1 1 0 11-1.414-1.414l.707-.707a1 1 0 011.414 0zM22 12a1 1 0 01-1 1h-1a1 1 0 110-2h1a1 1 0 011 1zM5 12a1 1 0 01-1 1H3a1 1 0 110-2h1a1 1 0 011 1zm14.66 8.66a1 1 0 01-1.414 0l-.707-.707a1 1 0 111.414-1.414l.707.707a1 1 0 010 1.414zM6.046 6.754a1 1 0 01-1.414 0l-.707-.707a1 1 0 011.414-1.414l.707.707a1 1 0 010 1.414zM12 7a5 5 0 100 10 5 5 0 000-10z"/></svg>`;
 
 /**
- * Pure helper: should we seed the DB with the OS-derived theme?
- *
- * We seed ONLY when all of:
- *   (a) there is no stored theme in the DB yet, AND
- *   (b) the localStorage→DB migration has already completed.
- *
- * The gate on (b) is critical: `initTheme()` runs at module-eval, before
- * `window.onload`. The migration runs later in `window.onload` and also
- * PATCHes the `theme` key (with the user's saved legacy value). If we seeded
- * unconditionally, the OS-derived write could land AFTER the migration write
- * and silently overwrite the user's saved theme (data loss). Deferring until
- * the migration flag is present means: if the flag is absent we skip the seed
- * this session — the migration will write the correct value, and a subsequent
- * load will reach the seed branch only on a genuinely fresh install.
- */
-export function shouldSeedTheme(hasStoredTheme: boolean, migrationDone: boolean): boolean {
-    return !hasStoredTheme && migrationDone;
-}
-
-/**
- * Applies the stored DB theme. Invoked from the boot sequence (`index.ts`
- * onload) AFTER the localStorage→DB migration completes, so the migrated theme
- * is already in the cache and applies on the FIRST post-upgrade load — not a
- * load later. On a genuinely fresh install (migration complete, no stored
- * theme), seeds the DB from the OS reading initTheme's first paint already set.
- *
- * The seed stays gated on `shouldSeedTheme` (migration-complete) as a
- * defensive guard: even though this now runs post-migration, the gate ensures
- * the OS-seed can never race the migration's own `theme` write.
+ * Applies the stored DB theme, invoked from the boot sequence (`index.ts`
+ * onload). On a fresh install (no stored theme) it seeds the DB from the OS
+ * reading that initTheme's synchronous first paint already applied, so the
+ * choice persists for the next load.
  */
 export async function applyStoredTheme(): Promise<void> {
     await settingsService.loadGlobal();
     const stored = settingsService.getGlobalCached()['theme'];
     if (stored === 'dark' || stored === 'light') {
         setTheme(stored);
-    } else if (shouldSeedTheme(false, localStorage.getItem(LEGACY_KEYS.migratedFlag) !== null)) {
-        // Fresh install (migration complete, no stored theme): persist the OS reading.
+    } else {
+        // Fresh install (no stored theme): persist the OS reading.
         void settingsService.patchGlobal({ theme: getTheme() }).catch(() => {});
     }
 }
 
 export function initTheme(): void {
     // Synchronous first paint from the OS preference — no await, no flash-to-blank.
-    // The stored (migrated) theme is applied later by applyStoredTheme(), invoked
-    // from the boot sequence AFTER the localStorage→DB migration runs in onload.
+    // The stored theme is applied later by applyStoredTheme() from the boot sequence.
     setTheme(firstPaintTheme(window.matchMedia('(prefers-color-scheme: dark)').matches));
 }
